Moved a couple of mapping classes so they are usable in the UI.
