+++ s-ramp-atom/src/main/java/org/overlord/sramp/atom/mappers/RdfToOntologyMapper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.overlord.sramp.server.atom.mappers;
+package org.overlord.sramp.atom.mappers;
 
 import java.net.URI;
 import java.util.ArrayList;
