+++ s-ramp-client/src/test/java/org/overlord/sramp/client/SrampAtomApiClientTest.java
@@ -39,6 +39,7 @@ import org.oasis_open.docs.s_ramp.ns.s_ramp_v1.XmlDocument;
 import org.oasis_open.docs.s_ramp.ns.s_ramp_v1.XsdDocument;
 import org.overlord.sramp.atom.archive.SrampArchive;
 import org.overlord.sramp.atom.err.SrampAtomException;
+import org.overlord.sramp.atom.mappers.RdfToOntologyMapper;
 import org.overlord.sramp.atom.providers.HttpResponseProvider;
 import org.overlord.sramp.atom.providers.SrampAtomExceptionProvider;
 import org.overlord.sramp.client.ontology.OntologySummary;
@@ -51,7 +52,6 @@ import org.overlord.sramp.common.test.resteasy.BaseResourceTest;
 import org.overlord.sramp.repository.PersistenceFactory;
 import org.overlord.sramp.repository.jcr.JCRRepositoryFactory;
 import org.overlord.sramp.repository.jcr.modeshape.JCRRepositoryCleaner;
-import org.overlord.sramp.server.atom.mappers.RdfToOntologyMapper;
 import org.overlord.sramp.server.atom.services.ArtifactResource;
 import org.overlord.sramp.server.atom.services.BatchResource;
 import org.overlord.sramp.server.atom.services.FeedResource;
