+++ s-ramp-server/src/main/java/org/overlord/sramp/server/atom/services/OntologyResource.java
@@ -35,11 +35,11 @@ import org.jboss.resteasy.plugins.providers.atom.Source;
 import org.overlord.sramp.atom.MediaType;
 import org.overlord.sramp.atom.SrampAtomUtils;
 import org.overlord.sramp.atom.err.SrampAtomException;
+import org.overlord.sramp.atom.mappers.OntologyToRdfMapper;
+import org.overlord.sramp.atom.mappers.RdfToOntologyMapper;
 import org.overlord.sramp.common.ontology.SrampOntology;
 import org.overlord.sramp.repository.PersistenceFactory;
 import org.overlord.sramp.repository.PersistenceManager;
-import org.overlord.sramp.server.atom.mappers.OntologyToRdfMapper;
-import org.overlord.sramp.server.atom.mappers.RdfToOntologyMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.w3._1999._02._22_rdf_syntax_ns_.RDF;
